requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
easyocr==1.4.1
fastapi==0.70.1
h11==0.12.0
httptools==0.3.0
idna==3.3
imageio==2.9.0
imutils==0.5.4
networkx==2.6.3
numpy==1.21.4
opencv-python
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.2.0
pydantic==1.8.2
pyparsing==3.0.6
pytesseract==0.3.8
python-bidi==0.4.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyWavelets==1.2.0
PyYAML==6.0
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
